Fixed warning.

+++ libdb/bt_debug.c
@@ -73,7 +73,7 @@ __bt_dump(dbp)
 	(void)fprintf(stderr, "%s: pgsz %d",
 	    F_ISSET(t, B_INMEM) ? "memory" : "disk", t->bt_psize);
 	if (F_ISSET(t, R_RECNO))
-		(void)fprintf(stderr, " keys %lu", t->bt_nrecs);
+		(void)fprintf(stderr, " keys %lu", (long unsigned int)t->bt_nrecs);
 #undef X
 #define	X(flag, name) \
 	if (F_ISSET(t, flag)) { \
@@ -112,12 +112,12 @@ __bt_dmpage(h)
 	char *sep;
 
 	m = (BTMETA *)h;
-	(void)fprintf(stderr, "magic %lx\n", m->magic);
-	(void)fprintf(stderr, "version %lu\n", m->version);
-	(void)fprintf(stderr, "psize %lu\n", m->psize);
-	(void)fprintf(stderr, "free %lu\n", m->free);
-	(void)fprintf(stderr, "nrecs %lu\n", m->nrecs);
-	(void)fprintf(stderr, "flags %lu", m->flags);
+	(void)fprintf(stderr, "magic %lx\n", (long unsigned int)m->magic);
+	(void)fprintf(stderr, "version %lu\n", (long unsigned int)m->version);
+	(void)fprintf(stderr, "psize %lu\n", (long unsigned int)m->psize);
+	(void)fprintf(stderr, "free %lu\n", (long unsigned int)m->free);
+	(void)fprintf(stderr, "nrecs %lu\n", (long unsigned int)m->nrecs);
+	(void)fprintf(stderr, "flags %lu", (long unsigned int)m->flags);
 #undef X
 #define	X(flag, name) \
 	if (m->flags & flag) { \
@@ -217,15 +217,15 @@ __bt_dpage(h)
 			if (bl->flags & P_BIGKEY)
 				(void)fprintf(stderr,
 				    "big key page %lu size %u/",
-				    *(pgno_t *)bl->bytes,
-				    *(u_int32_t *)(bl->bytes + sizeof(pgno_t)));
+				    (long unsigned int)*(pgno_t *)bl->bytes,
+				    (unsigned int)*(u_int32_t *)(bl->bytes + sizeof(pgno_t)));
 			else if (bl->ksize)
 				(void)fprintf(stderr, "%s/", bl->bytes);
 			if (bl->flags & P_BIGDATA)
 				(void)fprintf(stderr,
 				    "big data page %lu size %u",
-				    *(pgno_t *)(bl->bytes + bl->ksize),
-				    *(u_int32_t *)(bl->bytes + bl->ksize +
+				    (long unsigned int)*(pgno_t *)(bl->bytes + bl->ksize),
+				    (unsigned int)*(u_int32_t *)(bl->bytes + bl->ksize +
 				    sizeof(pgno_t)));
 			else if (bl->dsize)
 				(void)fprintf(stderr, "%.*s",
@@ -236,8 +236,8 @@ __bt_dpage(h)
 			if (rl->flags & P_BIGDATA)
 				(void)fprintf(stderr,
 				    "big data page %lu size %u",
-				    *(pgno_t *)rl->bytes,
-				    *(u_int32_t *)(rl->bytes + sizeof(pgno_t)));
+				    (long unsigned int)*(pgno_t *)rl->bytes,
+				    (unsigned int)*(u_int32_t *)(rl->bytes + sizeof(pgno_t)));
 			else if (rl->dsize)
 				(void)fprintf(stderr,
 				    "%.*s", (int)rl->dsize, rl->bytes);
@@ -306,10 +306,10 @@ __bt_stat(dbp)
 	(void)fprintf(stderr, "%d level%s with %ld keys",
 	    levels, levels == 1 ? "" : "s", nkeys);
 	if (F_ISSET(t, R_RECNO))
-		(void)fprintf(stderr, " (%ld header count)", t->bt_nrecs);
+		(void)fprintf(stderr, " (%ld header count)", (long int)t->bt_nrecs);
 	(void)fprintf(stderr,
 	    "\n%lu pages (leaf %ld, internal %ld, overflow %ld)\n",
-	    pinternal + pleaf + pcont, pleaf, pinternal, pcont);
+	    (long unsigned int)(pinternal + pleaf + pcont), (long unsigned int)pleaf, (long unsigned int)pinternal, (long unsigned int)pcont);
 	(void)fprintf(stderr, "%ld cache hits, %ld cache misses\n",
 	    bt_cache_hit, bt_cache_miss);
 	(void)fprintf(stderr, "%ld splits (%ld root splits, %ld sort splits)\n",
